requirements.txt
asn1crypto==0.24.0
cached-property==1.5.1
certifi==2018.4.16
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
future==0.16.0
idna==2.8
mock==2.0.0
pbr==5.1.3
peewee==3.5.0
peewee-migrate==1.1.5
pycparser==2.19
python-telegram-bot==11.1.0
redis==2.10.6
requests==2.21.0
six==1.12.0
urllib3==1.24.1
